remove invitationUrl field from request body
