+++ server/api-service/lowcoder-server/src/main/java/org/lowcoder/api/authentication/AuthenticationController.java
@@ -13,7 +13,6 @@ import org.lowcoder.api.usermanagement.view.APIKeyVO;
 import org.lowcoder.api.util.BusinessEventPublisher;
 import org.lowcoder.domain.authentication.FindAuthConfig;
 import org.lowcoder.domain.user.model.APIKey;
-import org.lowcoder.domain.user.service.EmailCommunicationService;
 import org.lowcoder.domain.user.service.UserService;
 import org.lowcoder.sdk.auth.AbstractAuthConfig;
 import org.lowcoder.sdk.util.CookieHelper;
@@ -36,7 +35,6 @@ public class AuthenticationController implements AuthenticationEndpoints
     private final CookieHelper cookieHelper;
     private final BusinessEventPublisher businessEventPublisher;
     private final UserService userService;
-    private final EmailCommunicationService emailCommunicationService;
 
     /**
      * login by email or phone with password; or register by email for now.
@@ -136,14 +134,5 @@ public class AuthenticationController implements AuthenticationEndpoints
     public Mono<ResponseView<?>> bindEmail(@RequestParam String email) {
         return sessionUserService.getVisitor().flatMap(user -> userService.bindEmail(user, email))
                 .map(ResponseView::success);
-    }
-
-    @Override
-    public Mono<ResponseView<Boolean>> sendInvitationEmails(InviteEmailRequest req) {
-        boolean isSuccess = emailCommunicationService.sendInvitationEmails(req.emails(), 
-        req.inviteLink(), 
-        "You have been invited to join our platform. Click here to accept the invitation: %s");
-        
-        return Mono.just(ResponseView.success(isSuccess));
     }
 }
