+++ server/api-service/lowcoder-server/src/main/java/org/lowcoder/api/authentication/AuthenticationEndpoints.java
@@ -1,7 +1,7 @@
 package org.lowcoder.api.authentication;
 
-import com.fasterxml.jackson.annotation.JsonView;
-import io.swagger.v3.oas.annotations.Operation;
+import java.util.List;
+
 import org.lowcoder.api.authentication.dto.APIKeyRequest;
 import org.lowcoder.api.authentication.dto.AuthConfigRequest;
 import org.lowcoder.api.framework.view.ResponseView;
@@ -13,11 +13,20 @@ import org.lowcoder.infra.constant.NewUrl;
 import org.lowcoder.sdk.auth.AbstractAuthConfig;
 import org.lowcoder.sdk.config.JsonViews;
 import org.lowcoder.sdk.constants.AuthSourceConstants;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ServerWebExchange;
-import reactor.core.publisher.Mono;
 
-import java.util.List;
+import com.fasterxml.jackson.annotation.JsonView;
+
+import io.swagger.v3.oas.annotations.Operation;
+import reactor.core.publisher.Mono;
 
 @RestController
 @RequestMapping(value = {NewUrl.CUSTOM_AUTH})
@@ -151,15 +160,6 @@ public interface AuthenticationEndpoints
 	)
 	@PostMapping("/email/bind")
 	public Mono<ResponseView<?>> bindEmail(@RequestParam String email);
-	
-	@Operation(
-			tags = TAG_AUTHENTICATION,
-			operationId = "sendInvitationEmails",
-			summary = "Send invitation emails",
-			description = "Send invitation emails to the specified addresses"
-	)
-	@PostMapping("/email/invite")
-	public Mono<ResponseView<Boolean>> sendInvitationEmails(@RequestBody InviteEmailRequest req);
 
     /**
      * @param loginId phone number or email for now.
@@ -168,11 +168,5 @@ public interface AuthenticationEndpoints
      */
     public record FormLoginRequest(String loginId, String password, boolean register, String source, String authId) {
     }
-	/**
-	 * @param emails email addresses to send the invitation to
-	 * @param inviteLink the link to be included in the email
-	 */
-	public record InviteEmailRequest(String[] emails, String inviteLink) {
-	}
 }
